+++ src/tests/Data-Layer/Models/YahooFinanceStock.test.ts
@@ -1,6 +1,6 @@
 const db = require("../../../Data-Layer/DataConnection");
-import StockModel from "../../../Models/Stock";
-import YahooFinanceStockModel from "../../../Models/YahooFinanceStock";
+import StockModel from "../../../Models/StockModel";
+import YahooFinanceStockModel from "../../../Models/YahooFinanceStockModel";
 
 beforeAll(() => (process.env.__DEV__ = "true"));
